solidify lb entrypoint

+++ loadbalancer/entrypoint.sh
@@ -2,34 +2,61 @@
 
 set -e
 
-fetch_config() 
-{
-    curl -s -f "$1" > haproxy.cfg
-    md5sum haproxy.cfg
-}
-
 scrape_config()
 {
+    # take the inital checksum
+    sum="$(md5sum haproxy.cfg)"
+
     while :
         do
-            sleep 60
-            old_sum="$(md5sum haproxy.cfg)"
-            sum="$(fetch_config "$1")"
-            if [ "$old_sum" = "$sum" ]
+            # sleep until next iteration
+            sleep "$2"
+
+            # try to fetch the config
+            if ! curl -s -f "$1" > haproxy.cfg;
             then
-                echo "config has not changed"
-            else
-                echo "config has changed, reloading worker"
-                kill -s SIGUSR2 1
+                cp previous.cfg haproxy.cfg
+                continue
             fi
+
+            # compare check sums
+            old_sum="$sum"
+            sum="$(md5sum haproxy.cfg)"
+            if [ "$old_sum" = "$sum" ]
+            then 
+                continue
+            fi
+
+            # test if file is valid
+            if ! haproxy -c -f haproxy.cfg;
+            then
+                cp previous.cfg haproxy.cfg
+                sum="$old_sum"
+                continue
+            fi
+
+            # if file has been successfully fetched, 
+            # has a different checksum and is valid,
+            # reload the worker
+           kill -s SIGUSR2 1
         done
 }
 
-# run task in background to update config and restart proxy
-scrape_config "$MANAGER_ENDPOINT" &
+# make a backup
+cp haproxy.cfg previous.cfg
 
-# get inital config
-fetch_config "$MANAGER_ENDPOINT"
+# sleep a bit to wait for manager
+sleep 10
+
+# fetch the first config or use the backup
+curl -s -f "$MANAGER_ENDPOINT" > haproxy.cfg
+if ! haproxy -c -f haproxy.cfg;
+then
+    cp previous.cfg haproxy.cfg
+fi
+
+# run task in background  every minute to update config and restart if needed proxy
+scrape_config "$MANAGER_ENDPOINT" 60 &
 
 # exec original entrypoint to make it pid 1
 exec docker-entrypoint.sh haproxy -f /usr/local/etc/haproxy/haproxy.cfg
